+++ components/organisms/applications/ApplicationInfo.tsx
